Private

Package internals documentation.

Parsing

Parsing is currently part of the private API as the output types are liable to change. Once this has stabilised, this will move to the public API.

GraphQLParser.parse — Function

parse(str::AbstractString)

Parses a GraphQL executable document string.

Miscellaneous

GraphQLParser.ValidationError — Type

ValidationError

Abstract type for validation errors.

GraphQLParser.format_block_string — Method

format_block_string(str)

Perform the formatting defined in the specification.